moved updateParentsDoneStatus into DbOperation.kt
